+++ CHANGELOG.md
@@ -56,6 +56,7 @@ https://github.com/bitcrowd/sshkit.ex/compare/v0.0.3...v0.1.0
 * Fix error handling in `SSHKit.SSH.Channel.send/4` when sending stream data
 * Context properly handles the case where env is set to an empty map
 * Fix environment variables export in contexts with user, group, umask, path and env
+* Fix SCP upload not making any progress after receiving a warning [#62]
 
 ## `0.0.3` (2017-07-13)